requirements.txt
pyyaml==6.0.1
gpiozero==2.0
urllib3==2.1.0
requests==2.31.0
# telepot==12.7
blinkpy==0.19.0
passlib==1.7.4
pytest==8.0.0
RPi.GPIO==0.7.1
astral==3.2